sw: refactor printf function
