+++ sw/lib/src/timer.c
@@ -1,9 +1,8 @@
-// Copyright 2022 ETH Zurich and University of Bologna.
+// Copyright 2024 ETH Zurich and University of Bologna.
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 //
-// Nils Wistoff <nwistoff@iis.ee.ethz.ch>
-// Paul Scheffler <paulsc@iis.ee.ethz.ch>
+// Philippe Sauter <phsauter@iis.ee.ethz.ch>
 
 #include "timer.h"
 #include "util.h"
